requirements.txt

Django==5.2.3
django-extensions==4.1
djangorestframework==3.16.1
dnf==4.23.0




